homeworks/week_04/requirements.txt
pyarrow
fastparquet
pandas
google-cloud-storage